btf: Add spec types iterator
This commit introduces the TypesIterator which is used to iterate over
types of a given spec. Each type is copied before being set in the
iterator struct.

The iterator is very similar to the existing InstructionIterator, as
both export the Next() method and set the values and indices in the
public fields.

Suggested-by: Timo Beckers <timo@isovalent.com>
Signed-off-by: Martynas Pumputis <m@lambda.lt>